requirements.txt
gradio
typing
langchain
pdfplumber
tabula-py
sentence-transformers
pinecone-client
python-dotenv
google-generativeai>=0.8.3